requirements.txt
sqlalchemy
pandas
numpy
scikit-learn
pymysql
joblib
fastapi
pydantic
uvicorn
python-multipart
streamlit
requests
openai
python-dotenv
